feat: update maxLines to double everywhere
- updated Text, RichText, SelectableText, DefaultTextStyle

- updated helperMaxLines, errorMaxLines and hintMaxLines of InputDecoration

- updated RenderParagraph, TextPainter

- updated ParagraphStyle in flutter engine

+++ engine/src/flutter/lib/web_ui/lib/src/engine/skwasm/skwasm_impl/paragraph.dart
@@ -709,7 +709,7 @@ class SkwasmParagraphStyle implements ui.ParagraphStyle {
   SkwasmParagraphStyle({
     ui.TextAlign? textAlign,
     ui.TextDirection? textDirection,
-    int? maxLines,
+    double? maxLines,
     String? fontFamily,
     double? fontSize,
     double? height,
@@ -812,7 +812,7 @@ class SkwasmParagraphStyle implements ui.ParagraphStyle {
   final ui.TextDirection? _textDirection;
   final ui.FontWeight? _fontWeight;
   final ui.FontStyle? _fontStyle;
-  final int? _maxLines;
+  final double? _maxLines;
   final String? _fontFamily;
   final double? _fontSize;
   final double? _height;
